requirements.txt
fastapi>=0.105.0
uvicorn[standard]>=0.24.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9
passlib[bcrypt]>=1.7
bcrypt>=4.0.0
python-dotenv>=1.0
openai>=1.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27
pytest>=8.0
email-validator>=1.0
trafilatura>=1.12,<2
